+++ java/security/MessageDigest/sha1.hpp
@@ -38,14 +38,11 @@
 typedef struct {
     unsigned int H[5];
     unsigned int W[80];
-    long int lenW;
-    unsigned int sizeHi, sizeLo;
+    int lenW;
+    unsigned int sizeHi,sizeLo;
 } SHA_CTX;
 
 void SHA1_Init(SHA_CTX *ctx);
-
-void SHA1_Update(SHA_CTX *ctx, const void *dataIn, long int len);
-
+void SHA1_Update(SHA_CTX *ctx, const void *dataIn, int len);
 void SHA1_Final(unsigned char hashout[20], SHA_CTX *ctx);
-
 #endif //sha1_INCLUDED
